+++ .idea/modules.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/com.techie.backend.main.iml" filepath="$PROJECT_DIR$/.idea/modules/com.techie.backend.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/techie.iml" filepath="$PROJECT_DIR$/.idea/techie.iml" />
     </modules>
   </component>
