+++ .idea/modules/com.techie.backend.main.iml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="AdditionalModuleElements">
+    <content url="file://$MODULE_DIR$/../../back-end/build/generated">
+      <sourceFolder url="file://$MODULE_DIR$/../../back-end/build/generated" isTestSource="false" generated="true" />
+    </content>
+  </component>
+</module>
