+++ back-end/build.gradle
@@ -40,6 +40,11 @@ dependencies {
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.security:spring-security-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+	// youtube api v3 ------
+	implementation 'com.google.api-client:google-api-client:1.33.0'
+	implementation 'com.google.apis:google-api-services-youtube:v3-rev20230816-2.0.0'
+	implementation 'com.google.http-client:google-http-client-jackson2:1.39.2'
+	// ---------------------
 }
 
 def generated = 'build/generated'
