+++ back-end/src/main/java/com/techie/backend/global/config/RestClientConfig.java
@@ -1,0 +1,14 @@
+package com.techie.backend.global.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestClient;
+
+@Configuration
+public class RestClientConfig {
+
+    @Bean
+    public RestClient restClient() {
+        return RestClient.create();
+    }
+}
