+++ back-end/src/main/java/com/techie/backend/global/security/SecurityConfig.java
@@ -32,6 +32,7 @@ public class SecurityConfig {
                                 .requestMatchers("/api/users/**").permitAll()
                                 .requestMatchers("/static/**").permitAll()
                                 .requestMatchers("/log.png").permitAll()
+                                .requestMatchers("/videos/**").permitAll()
                                 .anyRequest().authenticated())
                 .formLogin((form) ->
                         form
