--- back-end/src/main/resources/application.yaml
@@ -1,31 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:mysql://localhost:3306/techie
-    username: root
-    password: 1234
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    hikari:
-      maximum-pool-size: 10
-      minimum-idle: 5
-      idle-timeout: 600000
-      max-lifetime: 1000000
-      pool-name: HikariCP
-
-  jpa:
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        format_sql: true
-        dialect: org.hibernate.dialect.MySQLDialect
-        use_sql_comments: true
-    show-sql: true
-
-  security:
-    jwt:
-      token:
-        key: {JWT Key}
-
-  sql:
-    init:
-      mode: always
