--- build.gradle
@@ -1,70 +1,0 @@
-plugins {
-	id 'java'
-	id 'org.springframework.boot' version '3.3.4'
-	id 'io.spring.dependency-management' version '1.1.6'
-}
-
-group = 'com.techie'
-version = '0.0.1-SNAPSHOT'
-
-java {
-	toolchain {
-		languageVersion = JavaLanguageVersion.of(17)
-	}
-}
-
-configurations {
-	compileOnly {
-		extendsFrom annotationProcessor
-	}
-	// https://stackoverflow.com/questions/76706612/standard-commons-logging-discovery-in-action-with-spring-jcl
-	configureEach {
-		exclude group: 'commons-logging', module: 'commons-logging'
-	}
-}
-
-repositories {
-	mavenCentral()
-}
-
-dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-data-jpa' // spring jpa
-	implementation 'org.springframework.boot:spring-boot-starter-security' // spring security
-	implementation 'org.springframework.boot:spring-boot-starter-web' // spring web
-	implementation group: 'org.modelmapper', name: 'modelmapper', version: '3.1.1' // modelmapper
-	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2'
-	implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
-	annotationProcessor "com.querydsl:querydsl-apt:${dependencyManagement.importedProperties['querydsl.version']}:jakarta"
-	annotationProcessor "jakarta.annotation:jakarta.annotation-api"
-	annotationProcessor "jakarta.persistence:jakarta.persistence-api"
-	compileOnly 'org.projectlombok:lombok'
-	developmentOnly 'org.springframework.boot:spring-boot-devtools'
-	runtimeOnly 'com.mysql:mysql-connector-j'
-	annotationProcessor 'org.projectlombok:lombok'
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
-	testImplementation 'org.springframework.security:spring-security-test'
-	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-	// youtube api v3 ------
-	implementation 'com.google.api-client:google-api-client:1.33.0'
-	implementation 'com.google.apis:google-api-services-youtube:v3-rev20230816-2.0.0'
-	implementation 'com.google.http-client:google-http-client-jackson2:1.39.2'
-	// ---------------------
-}
-
-def generated = 'build/generated'
-
-tasks.withType(JavaCompile) {
-	options.generatedSourceOutputDirectory = file(generated)
-}
-
-sourceSets {
-	main.java.srcDirs += "$projectDir/build/generated"
-}
-
-clean {
-	delete file('build/generated')
-}
-
-tasks.named('test') {
-	useJUnitPlatform()
-}
