--- src/main/java/com/techie/backend/global/config/AppConfig.java
@@ -1,38 +1,0 @@
-package com.techie.backend.global.config;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestClient;
-
-@Configuration
-public class AppConfig {
-
-    @Bean
-    public ModelMapper modelMapper() {
-        return new ModelMapper();
-    }
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .components(new Components())
-                .info(apiInfo());
-    }
-
-    @Bean
-    public Info apiInfo() {
-        return new Info()
-                .title("API")
-                .description("Swagger UI Test")
-                .version("1.0.0");
-    }
-
-    @Bean
-    public RestClient restClient() {
-        return RestClient.create();
-    }
-}
