--- src/main/java/com/techie/backend/global/exception/ExceptionType.java
@@ -1,21 +1,0 @@
-package com.techie.backend.global.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-public enum ExceptionType {
-
-    USER_NOT_FOUND(HttpStatus.NOT_FOUND, "사용자가 존재하지 않습니다."),
-    INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "유효하지 않은 토큰입니다."),
-    FORBIDDEN_ACTION(HttpStatus.FORBIDDEN, "권한이 없습니다.");
-
-
-    private final HttpStatus status;
-    private final String message;
-
-    ExceptionType(HttpStatus status, String message) {
-        this.status = status;
-        this.message = message;
-    }
-}
