--- src/main/java/com/techie/backend/global/exception/GlobalExceptionHandler.java
@@ -1,17 +1,0 @@
-package com.techie.backend.global.exception;
-
-import com.techie.backend.global.exception.user.UserNotFoundException;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Slf4j
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(UserNotFoundException.class)
-    public ResponseEntity<String> handleUserNotFoundException(UserNotFoundException ex) {
-        return new ResponseEntity<>(ex.getExceptionType().getMessage(), ex.getExceptionType().getStatus());
-    }
-}
