--- src/main/java/com/techie/backend/global/security/SecurityConfig.java
@@ -1,70 +1,0 @@
-package com.techie.backend.global.security;
-
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    @Bean
-    public BCryptPasswordEncoder encoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity, CustomUserDetailService userDetailService) throws Exception {
-        httpSecurity
-                .csrf(AbstractHttpConfigurer::disable)
-                .authorizeHttpRequests((authorize) ->
-                        authorize
-                                .requestMatchers("/v3/api-docs/**", "/swagger-ui/**", "/swagger-ui.html", "/swagger-resources/**", "/webjars/**").permitAll()
-                                .requestMatchers("/api/users/**").permitAll()
-                                .requestMatchers("/static/**").permitAll()
-                                .requestMatchers("/log.png").permitAll()
-                                .requestMatchers("/videos/**").permitAll()
-                                .anyRequest().authenticated())
-                .formLogin((form) ->
-                        form
-                                .usernameParameter("username")
-                                .passwordParameter("password")
-                                .loginProcessingUrl("/api/users/auth")
-                                .defaultSuccessUrl("/api/home", true)
-                                .permitAll()
-                                .failureForwardUrl("/api/users/authFail"))
-                .logout((logout) ->
-                        logout
-                                .logoutUrl("/api/users/logOut")
-                                .logoutSuccessUrl("/api/users/logOutSuccess")
-                                .clearAuthentication(true)
-                                .deleteCookies("JSESSIONID"))
-                .userDetailsService(userDetailService);
-
-        httpSecurity.addFilterBefore(corsFilter(), UsernamePasswordAuthenticationFilter.class);
-        return httpSecurity.build();
-    }
-
-    @Bean
-    public CorsFilter corsFilter() {
-        CorsConfiguration configuration = new CorsConfiguration();
-        configuration.setAllowCredentials(true);
-        configuration.addAllowedOriginPattern("http://localhost:3000");
-        configuration.addAllowedHeader("*");
-        configuration.addAllowedMethod("*");
-
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-
-        return new CorsFilter(source);
-    }
-}
