--- src/main/java/com/techie/backend/gpt_history/domain/GptHistory.java
@@ -1,21 +1,0 @@
-package com.techie.backend.gpt_history.domain;
-
-import com.techie.backend.user.entity.User;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@Getter
-@Entity
-@Table(name = "gpt_histories")
-public class GptHistory {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @JoinColumn(name = "user_id", nullable = false)
-    @ManyToOne(fetch = FetchType.LAZY)
-    private User user;
-}
