--- src/main/java/com/techie/backend/memo/domain/Memo.java
@@ -1,36 +1,0 @@
-package com.techie.backend.memo.domain;
-
-import com.techie.backend.user.entity.User;
-import com.techie.backend.video.domain.Video;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@NoArgsConstructor
-@Getter
-@Entity
-@Table(name  = "memo")
-public class Memo {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(columnDefinition = "TEXT", nullable = true)
-    private String content;
-
-    @Column(nullable = false)
-    private LocalDateTime createDate;
-
-    @Column(nullable = false)
-    private LocalDateTime modifyDate;
-
-    @JoinColumn(name = "user_id", nullable = false)
-    @ManyToOne(fetch = FetchType.LAZY)
-    private User user;
-
-    @JoinColumn(name = "video_id", nullable = true)
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Video video;
-}
