--- src/main/java/com/techie/backend/search_history/domain/SearchHistory.java
@@ -1,29 +1,0 @@
-package com.techie.backend.search_history.domain;
-
-import com.techie.backend.user.entity.User;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@NoArgsConstructor
-@Getter
-@Entity
-@Table(name = "search_histories")
-public class SearchHistory {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @JoinColumn(name = "user_id", nullable = false)
-    @ManyToOne(fetch = FetchType.LAZY)
-    private User user;
-
-    @Column(nullable = false)
-    private String query;
-
-    @Column(nullable = false)
-    private LocalDateTime searchDate;
-}
