--- src/main/java/com/techie/backend/user/repository/custom/UserRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.techie.backend.user.repository.custom;
-
-public interface UserRepositoryCustom {
-}
