--- src/main/java/com/techie/backend/user/repository/impl/UserRepositoryImpl.java
@@ -1,7 +1,0 @@
-package com.techie.backend.user.repository.impl;
-
-import com.techie.backend.user.repository.custom.UserRepositoryCustom;
-
-public class UserRepositoryImpl implements UserRepositoryCustom {
-
-}
