--- src/main/java/com/techie/backend/user/service/UserService.java
@@ -1,5 +1,0 @@
-package com.techie.backend.user.service;
-
-public interface UserService {
-
-}
