--- src/main/java/com/techie/backend/user/service/impl/UserServiceImpl.java
@@ -1,10 +1,0 @@
-package com.techie.backend.user.service.impl;
-
-import com.techie.backend.user.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class UserServiceImpl implements UserService {
-}
