--- src/main/java/com/techie/backend/video/controller/YoutubeController.java
@@ -1,21 +1,0 @@
-package com.techie.backend.video.controller;
-
-import com.techie.backend.video.domain.Category;
-import com.techie.backend.video.dto.VideoResponse;
-import com.techie.backend.video.service.YoutubeService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class YoutubeController {
-    private final YoutubeService youtubeService;
-
-    @GetMapping("/videos/{category}")
-    public String listBackendVideo(@PathVariable Category category) {
-        return youtubeService.listVideos(category);
-    }
-
-}
