--- src/main/java/com/techie/backend/video/domain/Category.java
@@ -1,7 +1,0 @@
-package com.techie.backend.video.domain;
-
-
-public enum Category {
-    CLOUD,
-    DATA
-}
