--- src/main/java/com/techie/backend/video/dto/VideoResponse.java
@@ -1,10 +1,0 @@
-package com.techie.backend.video.dto;
-
-import lombok.Data;
-
-@Data
-public class VideoResponse {
-    private Long id;
-    private String title;
-    private String description;
-}
