--- src/main/resources/db/data.sql
@@ -1,8 +1,0 @@
-INSERT INTO video (title, video_id, category)
-VALUES
-    ('클라우드 컴퓨팅 한 방 정리 [안될과학-긴급과학 X AWS]','exewHoMNjsQ', 'CLOUD'),
-    ('요즘 모르면 안 되는 클라우드 개념과 필수 서비스 무료 교육 알아보기!','eRTUmgODZtg', 'CLOUD'),
-    ('데이터 분석가가 되고 싶다면 이건 꼭 보고 결정 하세요', '6WVgLBzpM6s', 'DATA');
-
-
-
