--- src/test/java/com/techie/backend/BackendApplicationTests.java
@@ -1,13 +1,0 @@
-package com.techie.backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
